requirements.txt
streamlit
requests
google-generativeai
youtube-transcript-api